+++ CHANGELOG.md
@@ -84,6 +84,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2301 Split ITs jobs by roles
 - YLP-2311 Reorganize ITs
 - YLP-2351 migrate average daily dose stat and total insulin stat from stat.js to yourloops
+- YLP-2367 Patient list refacto
 
 ### Fixed
 - YLP-2145 Patient cannot see monitoring widgets when he's in a monitoring team and another non-monitoring team